enos: handle SLES guestregister.service unreliability
The SLES provided `guestregister.service` systemd unit is unreliable
enough that it will ~ 1/9 times when provisioning SLES instances. When
this happens the machine will never exec SUSEConnect and we'll get no
access to the SLES repositories and subsequently break our scenarios.

I will report this issue upstream, but until then we work around the
issue by always checking the `guestregister.service` state and
restarting the service if it has failed.

Signed-off-by: Ryan Cragun <me@ryan.ec>